fix(friendModel):adjust post and delete request method

+++ students/steven/Canchu/models/friendModel.js
@@ -46,20 +46,33 @@ module.exports = {
             })
         })
     },
-    deleteFriend: async(res,id) => {
-        const sql = 'DELETE FROM friendship WHERE id = ?'
-        db.query(sql, [id], (error, results) => {
+    deleteFriend: async(res,id,self_id) => {
+        const validate_sql = 'SELECT user_id FROM friendship WHERE id = ?'
+        db.query(validate_sql, [id], (error, results) => {
             if (error) {
                 console.error('Database error:', error);
                 return res.status(500).json({ error: 'Server error' });
             }
-            res.status(200).json({
-                data: {
-                    friendship: {
-                        id: id
+            if(results[0].user_id !== self_id){
+                console.error("要求刪除Request的ID為",self_id,"，但能同意的只有ID為",results[0].user_id,"的使用者，id為",id)
+                return res.status(400).json({ error: 'This user has no permission to agree friend request'})
+            } else {
+                console.log("通過。接下來...")
+                const sql = 'DELETE FROM friendship WHERE id = ?'
+                db.query(sql, [id], (error, results) => {
+                    if (error) {
+                        console.error('Database error:', error);
+                        return res.status(500).json({ error: 'Server error' });
                     }
-                }
-            });
+                    res.status(200).json({
+                        data: {
+                            friendship: {
+                                id: id
+                            }
+                        }
+                    });
+                })
+            }
         })
     },
     postAgree: async(res,id,self_id) => {
@@ -69,8 +82,8 @@ module.exports = {
                 console.error('Database error:', error);
                 return res.status(500).json({ error: 'Server error' });
             }
-            if(results.user_id !== self_id){
-                console.error("要求同意Request的ID為",self_id,"，但能同意的只有ID為",results.user_id,"的使用者")
+            if(results[0].user_id !== self_id){
+                console.error("要求同意Request的ID為",self_id,"，但能同意的只有ID為",results[0].user_id,"的使用者，id為",id)
                 return res.status(400).json({ error: 'This user has no permission to agree friend request'})
             } else {
                 console.log("通過。接下來...")
